// src/api/generated/models/instituteDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { InstituteDtoStatus } from './instituteDtoStatus';

export interface InstituteDto {
  id?: string;
  companyId?: string;
  name?: string;
  code?: string;
  status?: InstituteDtoStatus;
  createdAt?: string;
}